refactor: :recycle: move delete actor method up to home view
